+++ main.py
@@ -1,26 +1,24 @@
-def get_user_info(users_data: list) -> None:
-    for user in users_data:
-        print(f'Twój znajomy {user["name"]} z miejscowości {user["location"]} opublikował {user["posts"]} postów.')
+from utils.model import users
+from utils.controller import get_user_info, add_user, remove_user, update_user
 
 
-def add_user(users_data: list) -> None:
-    new_name: str = input('podaj imię nowego znajomego: ')
-    new_location: str = input('podaj lokalizację: ')
-    new_posts: str = input('podaj liczbę postów: ')
-    users_data.append({'name': new_name, 'location': new_location, 'posts': new_posts}, )
+def main():
+
+    while True:
+        print('============MENU=============')
+        print('0 - zakończ program')
+        print('1 - wyświetl co u znajomych')
+        print('2 - dodaj znajomego')
+        print('3 - usuń znajomego')
+        print('4 - zauktualizuj dane znajomego')
+        print('=============================')
+        choice: str = input('wybierz opcję MENU: ')
+        if choice == '0': break
+        if choice == '1': get_user_info(users)
+        if choice == '2': add_user(users)
+        if choice == '3': remove_user(users)
+        if choice == '4': update_user(users)
 
 
-def remove_user(users_data: list) -> None:
-    user_name: str = input('podaj imię znajomego do usunięcia: ')
-    for user in users_data:
-        if user['name'] == user_name:
-            users_data.remove(user)
-
-
-def update_user(users_data: list) -> None:
-    user_name: str = input('podaj imię użytkownika do aktualizacji: ')
-    for user in users_data:
-        if user['name'] == user_name:
-            user['name'] = input('podaj nowe imię użytkownika: ')
-            user['location'] = input('podaj nową lokalizację użytkownika: ')
-            user['posts'] = input('podaj liczbę postów: ')
+if __name__ == '__main__':
+    main()
